orb(reactotron): chore(docs): Fix links in various README files to point to our new docs site. (#1412)
Some of the  files had links that 404'd because of the recent
rearrangement of docs for the new documentation site.

I went through both the main README and the reactotron-app README files
and replaced all instances of direct links to docs with a link to the
corresponding docs.infinite.red/reactotron link.

I also removed a badge that was no longer working on one of the readmes. (23e0ed6fa4a9cdf422c1c523c1be0577ba249939)