Set up Algolia search (#47)
Sets up basic search for IR docs using Algolia.

## Future improvements

There are improvements I'd like to make in the future to make this
easier to use, primarily revolving around the fact that this site covers
documentation for several different tools

1. Standardize naming and casing ("Ignite-cli" in the top bar should
probably just be "Ignite", especially since there's a "CLI" subheading,
consistent capitalization on "React-native-mlkit", etc)
2. Segment results in the search modal by top-level product name (show
"Ignite", not "Boilerplate" for results from the boilerplate docs) OR
display breadcrumbs for each segment that show the entire header
hierarchy. Otherwise it's hard to tell which product results are for.
3. Get Algolia to prioritize search results for the product docs you are
currently viewing. I'm pretty sure this is possible with Algolia's facet
filter feature (if what we need isn't already indexed, we can add custom
values to facet by:
https://docsearch.algolia.com/docs/record-extractor#indexing-content-for-faceting).